refactor: extract doltlite_branches.c (dolt_branches vtab)
Move the dolt_branches virtual table implementation into its own
translation unit. Export doltliteBranchesModule for registration from
doltliteBranchRegister. Wire PROLLY_OBJS, amalgamation, and lint.
